Refactored data dictionary wizard
